+++ src/torchpme/_utils.py
@@ -1,4 +1,4 @@
-from typing import Union
+from typing import Optional
 
 import torch
 
@@ -9,15 +9,17 @@ def _validate_parameters(
     positions: torch.Tensor,
     neighbor_indices: torch.Tensor,
     neighbor_distances: torch.Tensor,
-    smearing: Union[float, None],
-    periodic: Union[torch.Tensor, None] = None,
+    periodic: Optional[torch.Tensor] = None,
+    pair_mask: Optional[torch.Tensor] = None,
+    node_mask: Optional[torch.Tensor] = None,
+    kvectors: Optional[torch.Tensor] = None,
 ) -> None:
     dtype = positions.dtype
     device = positions.device
 
     # check shape, dtype and device of positions
-    num_atoms = len(positions)
-    if list(positions.shape) != [len(positions), 3]:
+    num_atoms = positions.shape[-2]
+    if list(positions.shape) != [num_atoms, 3]:
         raise ValueError(
             "`positions` must be a tensor with shape [n_atoms, 3], got tensor "
             f"with shape {list(positions.shape)}"
@@ -38,14 +40,6 @@ def _validate_parameters(
     if cell.device != device:
         raise ValueError(
             f"device of `cell` ({cell.device}) must be same as that of the `positions` class ({device})"
-        )
-
-    if smearing is not None and torch.equal(
-        cell.det(), torch.tensor(0.0, dtype=cell.dtype, device=cell.device)
-    ):
-        raise ValueError(
-            "provided `cell` has a determinant of 0 and therefore is not valid for "
-            "periodic calculation"
         )
 
     # check shape, dtype & device of `charges`
@@ -119,4 +113,60 @@ def _validate_parameters(
             raise ValueError(
                 f"device of `periodic` ({periodic.device}) must be same as that of "
                 f"the `positions` class ({device})"
+            )
+
+    if pair_mask is not None:
+        if pair_mask.shape != neighbor_indices[:, 0].shape:
+            raise ValueError(
+                "`pair_mask` must have the same shape as the number of neighbors, "
+                f"got tensor with shape {list(pair_mask.shape)} while the number of "
+                f"neighbors is {neighbor_indices.shape[0]}"
+            )
+
+        if pair_mask.device != device:
+            raise ValueError(
+                f"device of `pair_mask` ({pair_mask.device}) must be same as that "
+                f"of the `positions` class ({device})"
+            )
+
+        if pair_mask.dtype != torch.bool:
+            raise TypeError(
+                f"type of `pair_mask` ({pair_mask.dtype}) must be torch.bool"
+            )
+
+    if node_mask is not None:
+        if node_mask.shape != (num_atoms,):
+            raise ValueError(
+                "`node_mask` must have shape [n_atoms], got tensor with shape "
+                f"{list(node_mask.shape)} where n_atoms is {num_atoms}"
+            )
+
+        if node_mask.device != device:
+            raise ValueError(
+                f"device of `node_mask` ({node_mask.device}) must be same as that "
+                f"of the `positions` class ({device})"
+            )
+
+        if node_mask.dtype != torch.bool:
+            raise TypeError(
+                f"type of `node_mask` ({node_mask.dtype}) must be torch.bool"
+            )
+
+    if kvectors is not None:
+        if kvectors.shape[1] != 3:
+            raise ValueError(
+                "`kvectors` must be a tensor of shape [n_kvecs, 3], got "
+                f"tensor with shape {list(kvectors.shape)}"
+            )
+
+        if kvectors.device != device:
+            raise ValueError(
+                f"device of `kvectors` ({kvectors.device}) must be same as that of "
+                f"the `positions` class ({device})"
+            )
+
+        if kvectors.dtype != dtype:
+            raise TypeError(
+                f"type of `kvectors` ({kvectors.dtype}) must be same as that of the "
+                f"`positions` class ({dtype})"
             )