+++ src/torchpme/calculators/calculator.py
@@ -51,19 +51,24 @@ class Calculator(torch.nn.Module):
         charges: torch.Tensor,
         neighbor_indices: torch.Tensor,
         neighbor_distances: torch.Tensor,
+        pair_mask: Optional[torch.Tensor] = None,
     ) -> torch.Tensor:
         # Compute the pair potential terms V(r_ij) for each pair of atoms (i,j)
         # contained in the neighbor list
         with profiler.record_function("compute bare potential"):
             if self.potential.smearing is None:
                 if self.potential.exclusion_radius is None:
-                    potentials_bare = self.potential.from_dist(neighbor_distances)
-                else:
-                    potentials_bare = self.potential.from_dist(neighbor_distances) * (
-                        1 - self.potential.f_cutoff(neighbor_distances)
+                    potentials_bare = self.potential.from_dist(
+                        neighbor_distances, pair_mask
                     )
+                else:
+                    potentials_bare = self.potential.from_dist(
+                        neighbor_distances, pair_mask
+                    ) * (1 - self.potential.f_cutoff(neighbor_distances, pair_mask))
             else:
-                potentials_bare = self.potential.sr_from_dist(neighbor_distances)
+                potentials_bare = self.potential.sr_from_dist(
+                    neighbor_distances, pair_mask
+                )
 
         # Multiply the bare potential terms V(r_ij) with the corresponding charges
         # of ``atom j'' to obtain q_j*V(r_ij). Since each atom j can be a neighbor of
@@ -109,6 +114,9 @@ class Calculator(torch.nn.Module):
         neighbor_indices: torch.Tensor,
         neighbor_distances: torch.Tensor,
         periodic: Optional[torch.Tensor] = None,
+        node_mask: Optional[torch.Tensor] = None,
+        pair_mask: Optional[torch.Tensor] = None,
+        kvectors: Optional[torch.Tensor] = None,
     ):
         r"""
         Compute the potential "energy".
@@ -145,6 +153,12 @@ class Calculator(torch.nn.Module):
         :param periodic: optional torch.tensor of shape ``(3,)`` indicating which
             directions are periodic (True) and which are not (False). If not
             provided, full periodicity is assumed.
+        :param node_mask: Optional torch.tensor of shape ``(len(positions),)`` that
+            indicates which of the atoms are masked.
+        :param pair_mask: Optional torch.tensor containing a mask to be applied to the
+            result.
+        :param kvectors: Optional precomputed k-vectors to be used in the Fourier
+            space part of the calculation.
         """
         _validate_parameters(
             charges=charges,
@@ -152,8 +166,10 @@ class Calculator(torch.nn.Module):
             positions=positions,
             neighbor_indices=neighbor_indices,
             neighbor_distances=neighbor_distances,
-            smearing=self.potential.smearing,
             periodic=periodic,
+            pair_mask=pair_mask,
+            node_mask=node_mask,
+            kvectors=kvectors,
         )
 
         # Compute short-range (SR) part using a real space sum
@@ -161,6 +177,7 @@ class Calculator(torch.nn.Module):
             charges=charges,
             neighbor_indices=neighbor_indices,
             neighbor_distances=neighbor_distances,
+            pair_mask=pair_mask,
         )
 
         if self.potential.smearing is None:
@@ -171,6 +188,8 @@ class Calculator(torch.nn.Module):
             cell=cell,
             positions=positions,
             periodic=periodic,
+            kvectors=kvectors,
+            node_mask=node_mask,
         )
 
         return self.prefactor * (potential_sr + potential_lr)