+++ src/torchpme/potentials/combined.py
@@ -80,18 +80,24 @@ class CombinedPotential(Potential):
         else:
             self.register_buffer("weights", initial_weights)
 
-    def from_dist(self, dist: torch.Tensor) -> torch.Tensor:
-        potentials = [pot.from_dist(dist) for pot in self.potentials]
+    def from_dist(
+        self, dist: torch.Tensor, pair_mask: Optional[torch.Tensor] = None
+    ) -> torch.Tensor:
+        potentials = [pot.from_dist(dist, pair_mask) for pot in self.potentials]
         potentials = torch.stack(potentials, dim=-1)
         return torch.inner(self.weights, potentials)
 
-    def sr_from_dist(self, dist: torch.Tensor) -> torch.Tensor:
-        potentials = [pot.sr_from_dist(dist) for pot in self.potentials]
+    def sr_from_dist(
+        self, dist: torch.Tensor, pair_mask: Optional[torch.Tensor] = None
+    ) -> torch.Tensor:
+        potentials = [pot.sr_from_dist(dist, pair_mask) for pot in self.potentials]
         potentials = torch.stack(potentials, dim=-1)
         return torch.inner(self.weights, potentials)
 
-    def lr_from_dist(self, dist: torch.Tensor) -> torch.Tensor:
-        potentials = [pot.lr_from_dist(dist) for pot in self.potentials]
+    def lr_from_dist(
+        self, dist: torch.Tensor, pair_mask: Optional[torch.Tensor] = None
+    ) -> torch.Tensor:
+        potentials = [pot.lr_from_dist(dist, pair_mask) for pot in self.potentials]
         potentials = torch.stack(potentials, dim=-1)
         return torch.inner(self.weights, potentials)